Add test for all bonds and angles in Gromacs
Add tests to make sure the TPR parser reads all the possible bonds and
angles potentials that are available in Gromacs. Each potential is
tested separately to make debug easier.

Tabulated bonds and angles are ignored for now.

The files in testsuite/MDAnalysisTests/data/tprs/all_bonded describe a
system that is supposed to contain all the bonded interactions excepted
the tabulated ones.