Make some coordinate parser optional on Py3
The DCD parser is not compatible with python 3, this commit makes it
optional on that version of python. As the LAMPPS coordinate parser uses
the DCD parser, it is made optional as well.

The TRZ parser requires some small modifications to be compatible. It is
made optional until is can be tested.